requirements.txt
django==1.8
django-ckeditor==5.4.0
MySQL-python==1.2.5
image==1.5.17